Move the led types to the correct package and the already implemented pure functions to the algebra package
