+++ app/com/github/pedrovgs/dotto/core/types/morse.scala
@@ -1,5 +1,19 @@
 package com.github.pedrovgs.dotto.core.types
 
-class morse {
+sealed trait MorseSymbol {
+  val representation: String
 
+  override def toString = representation
+}
+
+case object Dot extends MorseSymbol {
+  override val representation = "."
+}
+
+case object Dash extends MorseSymbol {
+  override val representation = "-"
+}
+
+case object Space extends MorseSymbol {
+  override val representation = "/"
 }
