--- app/com/github/pedrovgs/dotto/morse/MorseModel.scala
@@ -1,19 +1,0 @@
-package com.github.pedrovgs.dotto.morse
-
-sealed trait MorseSymbol {
-  val representation: String
-
-  override def toString = representation
-}
-
-case object Dot extends MorseSymbol {
-  override val representation = "."
-}
-
-case object Dash extends MorseSymbol {
-  override val representation = "-"
-}
-
-case object Space extends MorseSymbol {
-  override val representation = "/"
-}
