+++ test/com/github/pedrovgs/dotto/core/algebra/MorseAlgebraSpec.scala
@@ -1,32 +1,33 @@
-package com.github.pedrovgs.dotto.morse
+package com.github.pedrovgs.dotto.core.algebra
 
 import com.github.pedrovgs.dotto.Resources
-import com.github.pedrovgs.dotto.morse.{Dot, Dash, Space}
+import com.github.pedrovgs.dotto.core.algebra.morse._
+import com.github.pedrovgs.dotto.core.types.{Dash, Dot, Space}
 import org.scalatest.FlatSpec
 
-class MorseTranslatorSpec extends FlatSpec with Resources {
+class MorseAlgebraSpec extends FlatSpec with Resources {
 
   it should "return an empty list if the input is an empty String" in {
-    val emptyTranslation = MorseTranslator.toMorse("")
+    val emptyTranslation = toMorse("")
 
     assert(emptyTranslation.isEmpty)
   }
 
   it should "translate A into .-" in {
-    val translatedA = MorseTranslator.toMorse("A")
+    val translatedA = toMorse("A")
 
     assert(translatedA == Seq(Seq(Dot, Dash)))
   }
 
   it should "be case insensitive" in {
-    val uppercaseTranslation = MorseTranslator.toMorse("A")
-    val lowercaseTranslation = MorseTranslator.toMorse("a")
+    val uppercaseTranslation = toMorse("A")
+    val lowercaseTranslation = toMorse("a")
 
     assert(uppercaseTranslation == lowercaseTranslation)
   }
 
   it should "translate an easy word from string to morse" in {
-    val translation = MorseTranslator.toMorse("Love")
+    val translation = toMorse("Love")
 
     assert(translation == Seq(Seq(Dot, Dash, Dot, Dot), Seq(Dash, Dash, Dash), Seq(Dot, Dot, Dot, Dash), Seq(Dot)))
   }
@@ -46,8 +47,8 @@ class MorseTranslatorSpec extends FlatSpec with Resources {
   it should "translate long strings" in {
     val originalText = readLowerCaseTextResource("/anyText.txt")
 
-    val translatedText = MorseTranslator.toMorse(originalText)
-    val text = MorseTranslator.fromMorse(translatedText)
+    val translatedText = toMorse(originalText)
+    val text = fromMorse(translatedText)
 
     assert(text == originalText)
   }
@@ -56,14 +57,14 @@ class MorseTranslatorSpec extends FlatSpec with Resources {
     val originalText = readLowerCaseTextResource("/anyTextWithNonSupportedChars.txt")
     val originalTextWithoutUnsupportedChars = readLowerCaseTextResource("/anyText.txt")
 
-    val translatedText = MorseTranslator.toMorse(originalText)
-    val text = MorseTranslator.fromMorse(translatedText)
+    val translatedText = toMorse(originalText)
+    val text = fromMorse(translatedText)
 
     assert(text == originalTextWithoutUnsupportedChars)
   }
 
   it should "return an empty sequence if the input is full of non supported chars" in {
-    val translatedText = MorseTranslator.toMorse(".%$&/")
+    val translatedText = toMorse(".%$&/")
 
     assert(translatedText.isEmpty)
   }
@@ -71,7 +72,7 @@ class MorseTranslatorSpec extends FlatSpec with Resources {
   it should "return an empty string if the input does not contains a valid sequence of morse symbols" in {
     val invalidMorseSentence = Seq(Seq(Dot, Dot, Dash, Dot, Dash, Dot, Dot), Seq(Dot, Dot, Dash, Dot, Dash, Dot, Dot))
 
-    val translatedString = MorseTranslator.fromMorse(invalidMorseSentence)
+    val translatedString = fromMorse(invalidMorseSentence)
 
     assert(translatedString.isEmpty)
   }
@@ -89,7 +90,7 @@ class MorseTranslatorSpec extends FlatSpec with Resources {
       Seq(Dot, Dot, Dash, Dot, Dash, Dot, Dot),
       Seq(Dot, Dot, Dash, Dot, Dash, Dot, Dot))
 
-    val translatedString = MorseTranslator.fromMorse(invalidMorseSentence)
+    val translatedString = fromMorse(invalidMorseSentence)
 
     assert(translatedString == "pv gs")
   }
@@ -105,7 +106,7 @@ class MorseTranslatorSpec extends FlatSpec with Resources {
   }
 
   private def toMorseString(text: String): String = {
-    MorseTranslator.toMorse(text).map(_.mkString).reduce(_ + " " + _)
+    toMorse(text).map(_.mkString).reduce(_ + " " + _)
   }
 
 }
